fix: merge duplicated name' widgets
